rename eep-data classes
