+++ web-server/src/server/json/eep-data-event.ts
@@ -1,4 +1,4 @@
-export default interface EepEvent {
+export default interface EepDataEvent {
   type: 'CompleteReset' | 'DataAdded' | 'DataChanged' | 'DataRemoved' | 'ListChanged';
   payload: undefined;
   eventCounter: number;
