+++ web-server/src/server/json/observers/state-observer.ts
@@ -1,5 +1,5 @@
 import { Socket } from 'socket.io';
-import JsonDataStore from '../json-data-reducer';
+import JsonDataStore from '../eep-data-reducer';
 
 export default interface StateObserver {
   onStateChange: (store: Readonly<JsonDataStore>) => void;
